support language query parameter in URL
If you visit an English Apple Music page for another region where the native language is not English, the URL will have "?l=en-US" at the end. This should pass the language query to the iTunes Lookup API, which will return the genre names in English but seemingly not the song titles for some reason. Thanks Apple